added tests

+++ src/main/kotlin/com/leijendary/spring/template/api/v1/data/SampleMessage.kt
@@ -5,6 +5,6 @@ import java.math.BigDecimal
 
 data class SampleMessage(
     val column1: String,
-    val column2: Int,
+    val column2: Long,
     val amount: BigDecimal?
 ) : LocalizedData<SampleTranslationMessage>()
